% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstratreg_se.R
\name{pstratreg_se}
\alias{pstratreg_se}
\title{Standard errors for pstratreg}
\usage{
pstratreg_se(x, r = 100)
}
\arguments{
\item{x}{An object of class \code{pstratreg}, such as the output of a call to \code{pstratreg()}}

\item{r}{Number of bootstrap replicates}
}
\value{
Data frame with estimates, standard errors, and confidence intervals for many quantities of interest.
}
\description{
Calculate standard errors by a nonparametric bootstrap, and 95% confidence intervals by a Normal approximation
}
\examples{
sim <- pstratreg_sim(n = 100)
result <- pstratreg(
     formula_y = formula(y ~ x + a),
     formula_m = formula(m ~ x + a),
     data = sim,
     treatment_name = "a"
)
se <- pstratreg_se(result, r = 10)
}
