% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstratreg_sim.R
\name{pstratreg_sim}
\alias{pstratreg_sim}
\title{Simulate Data}
\usage{
pstratreg_sim(n = 100)
}
\arguments{
\item{n}{Integer number of cases}
}
\value{
Simulated data frame
}
\description{
Simulate data to illustrate principal stratification regression bounding.
}
\examples{
pstratreg_sim(n = 100)

}
