% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pstratreg.R
\name{summary.pstratreg}
\alias{summary.pstratreg}
\title{Summary}
\usage{
\method{summary}{pstratreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pstratreg}, such as the output of a call to \code{pstratreg()}}

\item{...}{Other arguments to \code{summary} commands}
}
\description{
Summarize results from principal stratification regression.
}
