% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pstratreg.R
\name{print.pstratreg}
\alias{print.pstratreg}
\title{Print}
\usage{
\method{print}{pstratreg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pstratreg}, such as the output of a call to \code{pstratreg()}}

\item{...}{Other arguments to \code{print} commands}
}
\value{
Nothing. Prints summary of object
}
\description{
Print results from principal stratification regression.
}
