% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstratreg.R
\name{pstratreg}
\alias{pstratreg}
\title{Estimate Principal Stratification Regression Bounds}
\usage{
pstratreg(
  formula_y,
  formula_m,
  formula_sq_resid = NULL,
  family_y = "gaussian",
  homoskedastic = T,
  data,
  weights = NULL,
  treatment_name,
  monotonicity_positive = FALSE,
  monotonicity_negative = FALSE,
  aggregate = TRUE,
  group_vars = NULL
)
}
\arguments{
\item{formula_y}{A model formula for the outcome}

\item{formula_m}{A model formula for the binary indicator for whether the outcome exists}

\item{formula_sq_resid}{A model formula for the squared residuals for a variance function regression. Should only use variables that are used in \code{formula_y}. Because the Gamma regression for squared residuals is more computationally demanding, one might want a simpler \code{formula_sq_resid} than \code{formula_y}, perhaps to allow only highly structured heteroskedasticity. Defaults to match \code{formula_y}. No outcome need be specified in this formula.}

\item{family_y}{Character family of the outcome, either \code{"gaussian"} or \code{"binomial"}}

\item{homoskedastic}{A logical for whether homoskedasticity of the outcome should be assumed. If \code{FALSE}, estimates by variance function regression for log squared residuals with the same formula as in \code{formula_y}.}

\item{data}{A data frame}

\item{weights}{A numeric vector of weights, of length \code{nrow(data)}}

\item{treatment_name}{A character for the name of the treatment in \code{data},}

\item{monotonicity_positive}{A logical. Whether to assume M1 >= M0, so that treatment never causes the outcome to be undefined}

\item{monotonicity_negative}{A logical. Whether to assume M1 <= M0, so that lack of treatment never causes the outcome to be undefined}

\item{aggregate}{A logical. Whether to aggregate results or return one estimate per row of \code{data}}

\item{group_vars}{Character vector of names of variables in \code{data} by which to group when aggregating results. Only relevant when \code{aggregate = TRUE}}
}
\value{
An object of class \code{pstratreg}
}
\description{
Uses principal stratification and parametric models to bound the average causal effect among those who would have a valid outcome under either treatment condition
}
\examples{
sim <- pstratreg_sim(n = 100)
result <- pstratreg(
     formula_y = formula(y ~ x + a),
     formula_m = formula(m ~ x + a),
     data = sim,
     treatment_name = "a"
)
result <- pstratreg(
     formula_y = formula(y ~ x + a),
     formula_m = formula(m ~ x + a),
     data = sim,
     treatment_name = "a",
     aggregate = TRUE,
     homoskedastic = FALSE,
     monotonicity_negative = TRUE
)

}
